Code consolidation
